Added VsoUtil commands to update system configuration settings.

+++ src/Common/Src/VsoUtil/DeleteEnvironmentCommand.cs
@@ -53,16 +53,9 @@ namespace Microsoft.VsSaaS.Services.CloudEnvironments.VsoUtil
         /// <inheritdoc/>
         protected override void ExecuteCommand(IServiceProvider services, TextWriter stdout, TextWriter stderr)
         {
-            Guid id;
-
-            try
+            if (!Guid.TryParse(EnvironmentId, out var id))
             {
-                id = Guid.Parse(EnvironmentId);
-            }
-            catch
-            {
-                stderr.WriteLine($"Invalid Cloud Environment ID: {EnvironmentId}");
-                return;
+                throw new Exception($"Invalid Cloud Environment ID: {EnvironmentId}");
             }
 
             DeleteEnvironmentAsync(services, id, stdout, stderr).Wait();
@@ -85,8 +78,7 @@ namespace Microsoft.VsSaaS.Services.CloudEnvironments.VsoUtil
                 }
                 catch (Exception ex)
                 {
-                    await stderr.WriteLineAsync($"Environment not found: {EnvironmentId}. {ex.Message}");
-                    return;
+                    throw new Exception($"Environment not found: {EnvironmentId}. {ex.Message}", ex);
                 }
 
                 if (Verbose || DryRun)
